services/faces2_controllers/requirements.txt
paho-mqtt>=1.6.1
requests>=2.31.0
pydantic>=1.10.0